Patch r57 into the Beta branch.
r57: It turns out we were never properly removing the dragged tab from the source tabstrip when a drag was completed. This was causing various problems, which were highlighted when you reverted a drag (e.g. press esc or alt+tab while dragging).

QA has tested a build with this patch from trunk, and it looks good.


BUG=1262392
TBR=beng