Initial commit to get a handful of third_party packages building properly under Linux.  Uses some temporary shims to detect platform and omit processing of some stuff that doesn't seem to work for me.
As previously discussed, I'm just doing whatever's expedient to getting *something* building under Linux, since we decided with the current state of Chrome head (with the upcoming freeze and release branch) it would be too disruptive to try to land all the latest site_scons in Chrome's copy.  The main thing to note is the SConscript.main.linux file which uses

  #/../third_party/*

instead of

  #/$BUILD_ROOT/third_party/*

and then uses variant_dir in the env.SConscript() call at the bottom to explicitly emit output into Hammer/third_party/*.  If either of these hacks is absent, scons under Linux simply concludes that all targets are up to date (there's nothing to build).  Explanations for this behavior or alternative workarounds are welcome.  Alternatively we could just commit this as-is and let someone with more knowledge of SCons internals try to debug it under Linux.