Merge r286 into Beta to fix bug 1303133.
This fixes http://b/issue?id=1303133, which was a crash
in the renderer due to the plugin instance getting destroyed
in the context of the ShowModalHTMLDialog request. The fix
is to convert the WebPluginDelegateProxy object in a DeleteLater and to set its channel to NULL in the destructor. This ensures that the reply to the message can be sent out successfully.

Bug=1303133
TBR= iyengar@google.com