Merge r402 into Beta to fix a release blocker (crash on closing a tab).
r402:
Fix crash due to race conditions in the way Browser unhooks its
NotificationObserver for WebContents disconnection... not all codepaths would
result in Browser being removed as an observer. This change simplifies things a
bit by adding Browser as an observer for a TabContents whenever it is inserted
into the Browser's tabstrip, and removing it whenever it's detached. These
notifications are designed to be symmetrical and handle all use cases including
create/destroy(close) and also tab dragging/detaching.

Thanks to Adam for helping diagnose this and suggesting this fix.

B=1307678
TBR= beng@google.com