Rename HttpProxy* classes to Proxy*.  Move them into a net/proxy/ subdirectory.
I'm making this change because proxy resolution is really not specific to the HTTP protocol.  We need to use the proxy service in our FTP implementation, for example.  I made a separate directory instead of just putting these in base, because I anticipate more files once we have our own PAC implementation.