First, this removes the dependency of NavigationController on the SessionService. There were already some notifications that could be used to tell what's going on, and I filled out the rest. I moved some notifications from various places to the NavigationController.
In ssl_blocking_page, I removed a notification because it also calls
DidNavigateToEntry which does the notification (we got duplicate calls before).
In browser.cc I removed the tab parented notification since this is already
called by NavigationController::SetWindowID.

I removed some obsolete notifications and associated structures, especially in
history (like HOST_DELETED that was never issued).

I renamed the window_map_ in the MetricsService becuase it was actually a map
of controllers and Windows.

This also reorders the navigation_types header file to try to bring some order
to it.