Cleanup spell check file to stop using reference counted pointers without need.
I don't really believe this will solve anything (in terms
of the bug), but it did remove some needless complexity
above the crash site in the stack.  There is a tiny chance
that there is a shutdown race, where the destruction of the
containing object raced ahead of the execution (and
resulting destruction) of a task.  I think the destruction
of the task by the containing object would probably be
bad anyway, and I doubt that it even happens... so this
is all about being cleaner. 
bug=1318039

r=brettw