Fixed bug #1251296
This is a regression that is causing me to wonder why it ever worked in
the first place. The underflow bug was getting tripped by a very small scaling
matrix being improperly treated as a zero matrix, and also, the scaling code
was getting bitten by a bug in Skia's edge comparison function which caused
incorrect results to be returned when the difference between two values
exceeded the maximum signed integer.