Fixing issue 1343052: a regression where the Find box moves up when you open a new tab and close it again.
This is because during a TabClosingAt call, we end up calling ShowTabContents(NULL) which leads to ShelfVisibilityChangedImpl(NULL). Because the parameter current_tab is NULL, we end up removing the BookmarkBarView from the RootView (only to add it back a little too late), the Find window thinks there is no BookmarkBar and hence moves up a bit.

I am also enabling the disabled UI test I wrote for this, since it should now pass.

BUG=1343052
TEST=Covered by UI test.

Review URL: http://codereview.chromium.org/441