Fixes the following bugs reported with the Silverlight plugin:-
http://code.google.com/p/chromium/issues/detail?id=248
http://code.google.com/p/chromium/issues/detail?id=666

The basic issue here was that the plugin would not paint correctly.
The URLs mentioned in this bug load windowed silverlight plugin instances,
which invoke the NPN_InvalidateRect API to paint. We send over the rects
to the renderer, however these don't generate paints as the plugin is
windowed. A peek at Safari's webkit implementation revealed that they
merely invoke the InvalidateRect windows API in this context.

The fix is to emulate this behavior.


Review URL: http://codereview.chromium.org/431