Fixing bug 1951: Browser crash on View-Source of an https page. http://code.google.com/p/chromium/issues/detail?id=1951
We load the page in view-source mode, then request the favicon to show in the tab. The SSLManager gets notified about OnRequestStarted for both the page and the favicon, but the favicon request gets redirected from:

https://www.solomobile.ca/favicon.ico
... to ...
http://www.solomobile.ca/Error.aspx?aspxerrorpath=/Page-Not-Found.aspx

The SSLManager sees this as mixed content and while handling that it writes a debug message to the console, but in the process of writing the debug message we crash when GetTabContents(TAB_CONTENTS_WEB) returns NULL (since the current tab is TAB_CONTENTS_VIEW_SOURCE). Given that this is a debug message we are trying to write with a function that is only implemented on WebContents we should just NULL check the GetTabContents(TAB_CONTENTS_WEB) return value (and not log the debug message).

Review URL: http://codereview.chromium.org/2445