Fix for issue 2362: On-demand update hang with "Checking for update..."
We have been restructuring the message loop code lately and the MessageLoop on the file thread is not dispatching messages as it was before. I have changed the file thread to start with type MessageLoop::TYPE_IO, which pumps messages in such a way that Google Update can communicate back to us. 

I'm not sure what the best way to test this is but I ran the UI tests and the unit tests and they all pass.
Review URL: http://codereview.chromium.org/2918