Merge r3013 to the release branch:
Adds an --incognito switch that can be used to launch Chrome in Incognito mode.
You can combine this with --app= to launch an app in Chrome that has no effect
on the user's profile.

Note that this is not the same thing as launching chrome.exe
--user-data-dir=/dev/null, which would give you an completely empty profile
instead of loading the user's profile as read-only.

BUG=1790,2012

Patch by Yarin Kaul <yarin.kaul@gmail.com>

TBR= maruel@google.com
Review URL: http://codereview.chromium.org/7055