Merge r3226 to the release branch:
Add a support to NPN_PostUrl[Notify] using files

Post data to be uploaded from a file. This can be handled in two ways.
1. Read entire file and send the contents as if it was a post data 
   specified in the argument
2. Send just the file details and read them in the browser at the
   time of sending the request.

Approach 2 is more efficient but complicated. Approach 1 has a major
drawback of sending potentially large data over two IPC hops.  
In a way 'large data over IPC' problem exists as it is in case of 
plugin giving the data directly instead of in a file. Currently we
are going with the approach 1 to get the feature working. We can 
optimize this later with approach 2.

BUG=629

TBR= joshia@google.com
Review URL: http://codereview.chromium.org/7102