Merge r3323 to the release branch:
Make plugin processes (except Gears) run at normal priority.

This fixes a lot of the spurious unresponsive plugin dialogs, and also makes 
it so that the browser process isn't hung waiting for a low priority plugin 
process when cpu usage is high.

I was able to simulate this by using 100% cpu in a test program.  Before this 
change, the dialog always comes up and a flash video never plays.  The browser
process is also hung.  With this change the video plays and I can interact with
the browser.

TBR= jam@chromium.org
Review URL: http://codereview.chromium.org/7131