Several fixes to compile the latest merge.
Changes:
1) MessagePort DOM Object changes.
MessagePort now gets bindings generated.
Signature of window->postMessage changed.

2) Function names for Add/Remove listeners changed
on the Document and EventTargetNode objects.

3) SecurityOrigin changed upstream and no longer
has a securityToken() generator, but instead intends
for the caller to check if the SecurityOrigin was 
set in the DOM.

4) HTMLInputElement now offers custom getters for the
selectionStart and selectionStart.  This allows us to
eliminate our pseudo HTMLSelectionInputElement object.

5) SVGElementInstance added Add/RemoveListener functions.

TBR=fqian

This is not 100% working; dropping in so that others can
continue work on the merge.

Review URL: http://codereview.chromium.org/8130