Part 2: Getting release in synch with latest translations.
I need to port r3512 (tony) to pick up critical changes for the About box:
"""
The translation lead requested that this sentence be a single
string, otherwise it will likely be grammatically incorrect.

This is a hack until http://crbug.com/3295 is implemented.
"""
Once this is in, I'll need to replace the internal generated_resources.grd
with one from the release branch that does not include intermediate changes
from Brett (zoom) and Scott (context menu, bookmarks). I'm presuming I can
do that just in my local copy and don't have to check it into the internal
VCS.


TBR= tony@chromium.org
Review URL: http://codereview.chromium.org/8612