[Backport r4114 from the trunk.]
Pass a user agent string to WinHttpOpen to set the user
agent string on the session handle, which is the only
way to specify the User-Agent header for HTTP CONNECT
requests.

This requires some refactoring.  Most of the work done
by Session::Init() is moved to the Session constructor.
Session::Init() now only creates the primary session
handle, and the call to Session::Init() is delayed to
the first time we use the primary session handle
(session_->internet()).

R=mal
BUG=552
Review URL: http://codereview.chromium.org/8726