This fixes the VerifyHistoryLength ui test flakiness. The test initiates page navigations in timers and reads the history length to validate it.
When we receive a request to initiate a navigation in the browser, we send out the request to the renderer and then immediately read the history length and send it out in an update history length request. This causes the test to fail at times as it reads a stale history length. 

When we receive the DidAddHistoryItem notification in the renderer, we should not update the history length for the start page navigation. This results in the test failure at times as it reads an incorrect history length.

R=jam

Review URL: http://codereview.chromium.org/8898