Merge r3934, r3963, and r3971 to the release branch:
r3934: Make IPC::SyncChannel not duplicate the underlying  MessageLoop 
implementation by pumping messages on its own.  This fixes the problem of 
windowless plugins not painting on right click, and generally makes this 
class almost ported, other than using a generic version of events/locks.
Through this change I've also cleaned up the class and hopefully made it 
more understandable.

r3963: Fix some leaks in the syncchannel unittest.

r3971: Fix a problem with the SyncChannel rewrite.  DispatchReplies needs to
be called after the deserializer has been popped.  This now happens in the 
listener thread, so post a task to the IPC thread to do it.


BUG= 1432077, 1378101
TBR= jam@chromium.org
Review URL: http://codereview.chromium.org/9003