Merge r3568 to the release branch:
Remove throttling code from the Browser process and implement throttling in 
the Renderer.

The previous way of doing throttling was just calling CloseContents() on a 
window to reject it. But since the Browser is notified about a window 
opening asynchronously, by the time the CloseContents() sends a message back 
to the Renderer, a bunch more windows have been opened, leading to memory 
exhaustion. 

Instead, make all RenderViews created from a parent RenderView share a 
counter and start refusing to create RenderViews if too many RV have been 
created. Every RenderView (except for the first one) is assumed to be an 
unrequested popup, until notified by the Browser process by either a 
ViewMsg_DisassociateFromPopupCount message (this RenderView is a new top 
level page) or a ViewMsg_DisassociatePopup message (this RenderView is a 
requested popup and therefore shouldn't count against the count.)

BUG=3382, 2632
TBR= erg@google.com

Review URL: http://codereview.chromium.org/9205