Merge r3543 to the release branch:
Adds the anti-carpet bombing dialog. More specifically a new
EventHandler now exists between the buffered event handler and
download event handler. This new event handler asks the
DownloadRequestManager whether the download is allowed. This may
prompt the user and then the download continues or is canceled. The
DownloadRequestManager receives the request on the IO thread, forwards
to the UI thread, makes the decision, then notifies back on the IO
thread.

BUG=3422
TEST=make sure you don't see any problems downloading content.

TBR=sky@google.com

Review URL: http://codereview.chromium.org/8215