Merge r4005 to the release branch:
Reshuffling the options in the UnderTheHood section. A summary of my 
changes:Renamed ContentSection to WebContentSection to reflect changes 
that have already been made to the UI (prior to this change list).
Renamed GeneralSection to PrivacySectionMoved various checkboxes that 
appear in other categories in the "Under the hood"-tab to appear under 
the new PrivacySection.Moved the checkbox "Use a suggestion service" 
from Edit Search Engines to the PrivacySection, since we don't want it 
in two different places. - If you check the "Use Suggest" but have a 
provider that doesn't support it, it will just not perform Suggest
queries.The mock I was following has the Learn More link right after 
the text blurb, but this version has the link below the text. We need 
to figure out what we want to do here.

TBR= finnur@chromium.org
Review URL: http://codereview.chromium.org/9284