Merge r4050 to the release branch:
NOTE: Changes in web_contents_view_win.cc have to be made to 
web_contents.cc (Brett's changes to create _view_win.cc are not on
the release branch).

Make it so the user can't drag javascript: links to the
desktop from web content or the bookmarks bar.  Maintain
the ability to drag javascript: links from web content
to the bookmarks bar or to drag the javascript: link as
plain text.

BUG=3431
TBR= tony@chromium.org
Review URL: http://codereview.chromium.org/9300