Merge r4140 to the release branch:
Add option to disable/enable Spell Check. In addition, remove 
additional code in resource message filter, which was initilaizing 
spellchecker in the filter if it was NULL. This part of the code is
 not required since the resource message filter should not 
initialize spellchecker; it is up to the profile to give the 
spellchecker to the resource message filter.

Issue = 120

TBR= sidchat@chromium.org
Review URL: http://codereview.chromium.org/9096