Merge r4250 to the release branch:
Fix for HTTP AUTH prompts allowing save-password infobars to be shown after a transition away from the appropriate page while that page was loading. 

The issue was the LoginPrompt set up a provisional password save prior to the start of a page load, and DidNavigate is not called early enough when a user navigates, which they have an opportunity to do in the HTTP auth case (details in bug comments).

BUG=2792

TBR=glen@ghromium.org

Review URL: http://codereview.chromium.org/9123