Merge r4569 to the releae branch:
A fix for Issue 2768 "IME: Current clause of Japanese IME doesn't 
appear to reflect its rage change."
This issue is caused by my code that cannot extract a target clause 
from a composition string when there is an input clause after a 
target clause.
when a user changes the range of a target clause, Japanese IMEs 
creates a composite clause which consists of a new target clause and an input clause. My IME code treats such composite clause as a target clause and cannot reflect the change.
To fix this problem, this code changes the algorithm which extracts 
a target clause.
TBR= hbono@chromium.org
Review URL: http://codereview.chromium.org/9410