Merge r4636 to the release branch:
Fix hang in clearing browsing data from incognito mode.

When deleting Keywords, the BrowsingDataRemover class may need to 
load the TemplateURLModel if it isn't loaded and delete keywords 
asynchronously from the Observe function. The check to see if we 
were getting the model with the right profile wasn't taking into 
account that we might be doing this while incognito, which led to 
us failing to set the waiting_for_keywords_ flag, resulting in a 
hang (issue 2145). This fixes that.

TBR= finnur@chromium.org
Review URL: http://codereview.chromium.org/9411