Fix Google Chrome not launching.
r4284 consolidated constants, but did not inlcude the trailing \\
on a couple of paths.

As a result, chrome.exe will not run when chrome.dll is in a 
versioned subdirectory.

R= kuchhal
TEST= None. Sorry for the AOL caps, but WE REALLY NEED A TEST HERE.
This surfaced only after installing an official build. In most 
cases, developers will have chrome.exe and chrome.dll in the same 
dir and never find regressions here.

Review URL: http://codereview.chromium.org/9632