Merging missing file from r4958
r4958 did not pick up render_view_host.cc because of svn decided to act funky. So checking in the merged file in this revision.

r4935:
Continue with crosssite navigations if alerts are suppressed in an unload handler (i.e. we have an infinite-ish alert loop).

TBR=mal@chromium.org

Review URL: http://codereview.chromium.org/9678