Merge r5113 to the release branch:
Ensure that the textarea's shadow node does not collapse if it is empty.

I have a layout test for this as well (see the bug) that I'll be
committing upstream with webkit and a layout test in the chrome tree
that hasn't been upstreamed yet that I'll update when I update DEPS
to pull this in.

BUG=2033
TBR=ojan
Review URL: http://codereview.chromium.org/10742