Fix the set of fonts used for test_shell.
With this change, the Google homepage renders pixel exact to Firefox on my
Goobuntu box (which isn't using subpixel rendering). Note that, although the
pixels are the same, the colours are very slightly different when it comes to
anti-aliasing. I suspect this is because Skia is currently missing a gamma
table.

Review URL: http://codereview.chromium.org/11211