Adds various docking positions when dragging around a tab. I've added the ability to dock to the following locations:
Maximize the window on drop.

Resize the window to fill up the left, bottom, or right half of the
monitor.

Tile the newly created window with an existing chrome window such that
the two are on top of each other and fill the height of the monitor,
or tiled such that they fill the width of the monitor.

The graphics suxor. Glen says he'll come up with something better once
its landed.

BUG=none
TEST=4628

Review URL: http://codereview.chromium.org/11325