Merge 5927 - Don't reenter Webkit, while it waits for outgoing sync calls to complete. This fixes bug http://code.google.com/p/chromium/issues/detail?id=4582, which is a crash in the renderer process, which occurs when the renderer receives a sync message requesting accessibility information from the browser, while it waits for the PluginDestroyed sync call to unwind.
The ViewMsg_GetAccessibilityInfo has a timeout associated with it, which indicates that it is ok for the call to fail. We now turn off the unblock flag on the sync message which will ensure that it does not reenter Webkit in this scenario.

R=jcampan
Bug=4582
TBR=ananta@chromium.org

Review URL: http://codereview.chromium.org/12402
Review URL: http://codereview.chromium.org/12693