Local fix to SCons bug:
If a Node exists explicitly in a repository directory but not in the
local working directory, SCons would throw an exception when globbing
the entry when it tried to extract the non-existent same-named entry
from the local working directory's entries dictionary.  Since we know
we want to return local entry anyway, make sure it exists.
Review URL: http://codereview.chromium.org/13044