Added line number and source URL information to inspector console messages.
This uses a debugger API in V8 which can look at the JavaScript execution stack and find the line number and source URL for the top stack frame. The JavaScript functions required for this are added to a new utility context created in V8Proxy. This context is created as a singleton (on demand) as it has no dependency in the page currently running.

I will not submit this change until V8 version 0.4.5 is used in Chrome as this change depends on a new V8 debugger API added to that version.

The change also requires a second step to udate the WebKit DEPS.

BUG=2960
Review URL: http://codereview.chromium.org/12804