fix box shadows
We need to apply the context transform to the shadow before
rendering it.  There's also a weird case where we have to
flip the y offset if we're not applying the transform.  I'm
not sure why this is necessary, but it allows us to pass
canvas-shadow tests.

BUG=5503

Review URL: http://codereview.chromium.org/14125