* Launch Chrome in non-elevated mode.
Currently Google Update LaunchCmdLine launches command with the callers token. To launch non-elevated we try to get the token of explorer.exe and impersonate it (same as what Google Update does in another place).

Also fix a bug in Compatibility check function. We should exit if we fail to load advapi32.dll and not if we succeed.

I am still doing testing with various builds of Google Update.

Review URL: http://codereview.chromium.org/12970