Fix 5561: Support system-level install in the channel changer.
When Chrome is installed system wide, the Google Update keys reside under HKLM.
Under Vista, any writes to this key will get virtualized, unless we elevate
the channel_changer process. We don't want to elevate using a manifest, because
we don't want the UAC to pop up when the Channel Changer is used to configure a
user-level Chrome installation, so we detect at process startup whether we need
to elevate and relaunch the channel_changer process (elevated) if we need to
(before showing any UI).

Review URL: http://codereview.chromium.org/14855