Reverting 7296.
From http://build.chromium.org/buildbot/waterfall/waterfall?builder=Chromium%20Reliability
There is a clear increase of crashes after this change. It really seems like a heap overflow but it's hard to tell. Reverting to see if it helps the reliability tests. If not, we can reapply this change and find another culprit.
Review URL: http://codereview.chromium.org/15081