Add chrome-user-script:// protocol
This is a step towards getting user scripts working in extensions. It's a bit
janky to use the form chrome-user-script://<script file>/ (with no path), but
GURL assumes that there is always a host, but path is optional, making this
approach simpler than alternatives.

Review URL: http://codereview.chromium.org/16592