Report name of repeatedly registered histograms
Histograms should ONLY be listed once in the code
for a given process (the browser process?).  If 
two or more histograms have the same name, it is an
error, and a DCHECK will fire.  This code will also
display the name of the offending histogram.

r=sky
Review URL: http://codereview.chromium.org/16609