remove libxslt_config.vcproj and add generate header script
It turns out that config.h is not used on windows so we don't need
to check it in.  There's a libxslt/xsltconfig.h that is generated
and used, but it's been checked into the tree since initial.commit.

I also added a libxslt.vsprops to make updating include paths easier.

Review URL: http://codereview.chromium.org/18121