Merge r7669 into 154 branch:
When there are multiple requests for the same resource, it is possible that cancelling a request that is currently a reader may be racing with another reader being completed.  In that case, we were not removing the transaction for the cancelled request so all queued requests were blocked forever. 

BUG=4769
TEST=unittest

Review URL: http://codereview.chromium.org/18469