Separate out the backing store from the RenderWidgetHost into its own file to make porting easier.
I also did some cleanup in the scrolling and painting areas, collapsing the
confusing ScrollRect call inside OnMsgScrollRect to just be inside the function
and call a new function on the backing store to scroll it. Same for painting.

This also moves plugin window moving to the view.
Review URL: http://codereview.chromium.org/18702