Baselining 2 tests: onload-navigation-iframe.html onload-navigation-iframe-timeout.html
The tests submit a form to about:blank.
They fail because there is no ? at the end of about:blank when we land to the new page.
This is because the GURL implementation treats about:blank as a path URL and sets no query for it.

Note: the test diff files are in my home dir under the layoutest_onload dir.


Review URL: http://codereview.chromium.org/19492