Clean up WebKit resources after starting a download.
Downloads initiated in the renderer process have associated
resources that need to be cleaned up properly. Since we handle
downloads in the browser process, the renderer was not receiving
the completion notification and thus hanging on to resources.

We now send the renderer a completion notification once we detect
that a resource request is a download, so that the renderer can
properly clean up.

BUG=6531 (http://crbug.com/6531)

Review URL: http://codereview.chromium.org/20524