We can get into a state where the automation framework sends a search string, which ends up in a handler in the FindBarView that has already been destroyed, causing a crash.
We now explicitly null the controller for the text field when destroying the view so that if it is still receiving messages they will be ignored by the view.

(I can't reproduce this bug manually or by running the automation testing locally, but this is an attempt to fix the crash by using the information gathered from looking at the crash dump).

Review URL: http://codereview.chromium.org/27172