Linux: fix XRENDER support for NX
The NX X server doesn't support xRGB32 XRENDER picture formats. So, if
we don't find one of those formats, fall back on ARGB32, which is
fine, except that the X server will spend time dealing with an alpha
channel which is all 254 or 255. ARGB32 support is required by the
XRENDER spec.

(This doesn't fix VNC, a patch for that is forthcoming.)

Review URL: http://codereview.chromium.org/28192