Fix toolbar keyboard focus (shift-alt-t), which was broken by me, and toolbar button context menus on VK_APPS, which was broken by Jonas.
The overall focus issue was caused by my change that made View::RequestFocus() sanity-check that the View was focusable.  The toolbar uses a crazy hack where it purposefully wants to get focus even though it's not focusable.  :P

The context menu issue was caused by Jonas changing the name of a virtual function, presumably not realizing it was a virtual, not just a simple accessor.  I changed the name back and marked this function (and several others) as virtual.  In order to avoid blowing up the source in toolbar_view.cc, I reverted users of the accessor to just using the member variable name.

Also reordered a couple functions to match the order they were originally declared in View.h.
Review URL: http://codereview.chromium.org/27175