Enable --single-process for Mac and Linux.
Additionally, wire in a bit of setup for the --single-process case,
because it cannot be done by the renderer thread because that thread
is not the main thread.

Fair warning: --single-process often seems to be broken by various
unrelated changes, which may-or-may-not make assumptions about the
process architecture.  I will try to stay on top of these.

Review URL: http://codereview.chromium.org/21330