Pipeline_Impl was modified to properly render a stream that has video but no audio.  A unit test accompanies the change.
Note that one minor other change was snuck in with this change.  The data_source_impl.cc had a TODO to set a more specific error
when a read failed.  Because I was already updating the pipeline error enum, I added the error code, changed the call to
host_->Error(), and removed the TODO.
Review URL: http://codereview.chromium.org/39170