Merge r11285.
Fixes crash in ResizeCorner. ResizeCorner was caching the BrowserView
that created it, but if a tab contained a download shelf and was
dragged to a new window the ResizeCorner would be referencing the
wrong BrowserView. I've changed the code to look up the ancestor
BrowserView as necessary.

Review URL: http://codereview.chromium.org/41010
BUG=8477
TBR= sky

Review URL: http://codereview.chromium.org/42058