Merge r11424
Fix the distributor logo placement on the glass frame.

The window's control bounds can change independently of window sizing messages (WM_SIZE), and so we need to make sure that we always call Layout on the frame view and the client view, even if the bounds of the window don't change so that we are more likely to update the layout of the distributor logo.

BUG= 8511

TBR= ben
Review URL: http://codereview.chromium.org/42062