Basic windowless plugins, try 2.
In response to Dean's comment on http://codereview.chromium.org/39105,
I redid that patch to put the X munging into the plugin implementation.  This
won't work for multiproc, but it's near the correct place and many things will
need to be changed for multiproc.

Review URL: http://codereview.chromium.org/42056