Initial support for Linux build packaging and archiving.
There's still a lot to figure out and clean up, but this should get us going with automated packaging for dogfood pushes.

Note that we also still need a Linux buildbot with the build configuration we want archived. Once we have that, I'll have another change to enable archiving with this script.

Review URL: http://codereview.chromium.org/43122