Ensure that MenuGtks are destroyed before their accelerator group.
When a MenuGtk is created, it is passed a GtkAccelGroup.  Previously the GtkMenu widgets were destroyed after the browser window, which was the owner of the GtkAccelGroup.  Make the toolbar own a reference to the GtkAccelGroup, and teardown all GtkMenus that reference it beforehand.

Some MenuGtk cleanup while debugging the problem.

BUG=8866

Review URL: http://codereview.chromium.org/48065