Try to get inter-packet times more precisely
I had tried to rely on the call to ReadData() being made once per TCP/IP
packet, but histograms from the field show that does not work well :-(.
This version looks for boundary counts of multiples of 1430 bytes, and uses
that as a better approximation of when a packet is received.  For
test data with only a few packets, this is excellent.  Some vendors such
as Google tend to use at most 1430 bytes per packet, so even if we are
off a bit, it will be close (for small packet counts).

Alos, add histogram for total transmitted data (compressed via sdch and gzip)


r=huanr,ajenjo
Review URL: http://codereview.chromium.org/48077