Fully enable history_url_provider_unittest.cc for Mac.
gfx::GetCleanStringFromUrl() got implemented for Mac at some point (I
suspect when fonts landed), so now all of this code should be
portable.

There is a slight difference in how ElideUrl() works, due to font
metric differences, but that should not be relevant to this code.

Review URL: http://codereview.chromium.org/42513