Fix fast shutdown
The check for unload handlers was backwards.  This bug was introduced in r7901 <http://codereview.chromium.org/17453>.

Also, remove the call to process_.Close().  When we Terminate the process, we get a IPC channel error and BrowserRenderProcessHost::OnChannelError gets called.  This method needs the handle so it can call DidProcessCrash.  It also does the call to process_.Close() for us.

Review URL: http://codereview.chromium.org/42653