Temporary fix to avoid crash in typing in Omnibox. The crash happens if you have a lower case string that is longer than the mixed case string and we get a match at the end of the string. This crashed because the offset is now past the end of the string.
This fix is temporary in that it avoids the crash by not highlighting
a match in the string. The right fix will be better parsing so we
correctly recognize how the two strings line up, but that'll happen
later.

BUG=9335
TEST=Create a bookmark with the title on the page
http://www.google.com/support/forum/p/Chrome/thread?fid=3b64f0cdd1e29e94000466022c425763&hl=en
, then type in i and make sure we don't crash. This is also covered by
unit tests now, so don't feel you need to test it.

Review URL: http://codereview.chromium.org/42664