It looks like the Chrome NPAPI plugin installer has been broken since we first updated chrome webkit after 1.0 shipped.
Basically in the 1.0 branch when the plugin was instantiated in its instantiation it would get the mime type along with the list of other arguments. If an object tag was specified with the classid, it would get mapped to the mime type. With the webkit merge the classs id is passed in along with the mime type.

The plugin installer thinks that this is an activex installation on receiving a valid class id and 
and ends up checking if it is a white listed classid, etc.

All this code will be taken out along with the activex shim in the near future. For now we take this code path only if we don't have a valid mime type.

This fixes http://code.google.com/p/chromium/issues/detail?id=8584

Added a plugin test for the argument parsing functionality in the default plugin. I changed the ParseInstantiationArguments function in the
plugin installer to a static function to be able to unit test this.


Bug=8584

Review URL: http://codereview.chromium.org/42684