Remove some unused stuff from basictypes.h.
My original goal was just to remove down_cast, since it uses RTTI, and I am
going to update the build with -fno-rtti.  The other stuff I just noticed is
unrelated and unused in our code, so we might as well remove it, since it
hasn't proven to be useful.

Review URL: http://codereview.chromium.org/63121