Merge 13072 - A workaround fix for Issue 9186.
This crash is caused in a unnecessary check whether or not to add a "text direction" submenu to a context menu. Since we don't have a "text direction" submenu and this check is unnecessary now, this change set a page setting that disables this check.

BUG=9186
Review URL: http://codereview.chromium.org/42690
TBR=hbono@chromium.org
Review URL: http://codereview.chromium.org/67069