Add more prototype headers.  Tweak a method declaration on WebDragData to be consistent with similar methods in the new headers.
Renamed WebCursor to WebCursorInfo (to be like WebScreenInfo) and so that I can
keep WebCursor in webkit/glue as a "smart pointer" class for a native cursor
used to render the WebCursorInfo.  The changes to WebCursor are still pending;
no one is using WebCursorInfo yet.

Notice that I am using the PIMPL pattern for WebHistoryItem, WebURLRequest, and
WebURLResponse because in those cases the WebKit interface just wraps a
corresponding WebCore class that only contains data.  WebFrame, WebView, and
WebDataSource (among others) are vtables because they wrap more complex
implementation classes that we may wish to mock in unit tests.

I'm introducing WebPopupMenu as a more explicit way of referring to the
WebWidget used to contain a WebCore::PopupContainer.  I think this will
be slightly less confusing than the current situation where that is just
a WebWidget implemented via WebWidgetImpl.  Now instead we'll have
WebPopupMenuImpl.

Finally, I ditched stdint.h and the corresponding WIN32 typedefs.  It seems
that WebKit is moving away from these, so I should do the same.

R=dglazkov

Review URL: http://codereview.chromium.org/73073