Add concept of slow tests. Only mark one test as slow for now. After checkin, I'll add a ton of slow tests and reduce the timeout for non-slow tests to something like 1 second.
To aid in that, I've added 90th percentile and 99th percentile times to the statistics we report. Also made it so that tests marked as slow and timeout/crash tests are listed separately from the slowest 50 tests so that list can be used to know which tests to mark slow.

Tests cannot be marked as both SLOW and TIMEOUT. If the test times out, then we want it to timeout early to avoid waiting a long timeout for a test that we know is just going to timeout.

TestShell can now do per-test timeouts. We could easily make run-webkit-tests do per-test timeouts, but I think just having the one modifier of SLOW is easier to grok. Makes the test lists simpler and allows for having longer timeouts on debug/purify builders.

BUG=9324

Review URL: http://codereview.chromium.org/67198