Refactor DialogDelegate so the DialogButton enum is in cross platform code.
This is mostly a find & replace in our code to use the new location
of the enum.

I also deleted some GetDialogButtons methods in child classes
because they were identical to the parent version in DialogDelegate.

This will allow more code to be enabled on linux/mac in the
automation provider.

Review URL: http://codereview.chromium.org/88008