linux (and some posix): multiprocess plugins compiling.
The goal of this change is to *not* make any behavioral change, but to
instead just get all the plugin-related files linking on Linux with
a bunch of NOTIMPLEMENTED()s in the appropriate places.  It's enormous
enough already without any refactorings or new features.

Changes include:
 - Lots of gcc warning fixes.
 - Use portable replacements for Windows-specific functions (_strdup, etc.).
 - Use TransportDIB instead of just shared memory in the plugin messaging.
   Note that this is not fleshed out on Linux and on Windows it just hacks
   in the existing handles so there should be no functional change.
 - Fix --plugin-launcher to use cross-platform APIs.

Review URL: http://codereview.chromium.org/79020