Fix memory corruption caused by missing sizeof(T) in memory allocation code.
I also noticed that assign from a container C is missing a check for size() > 0
prior to asking for element 0.  This matches the code in the copying
constructor that takes a container C parameter.  This change is unrelated to
the bug at hand, but it seems like something that should also be fixed.

BUG=11045
R=dglazkov

Review URL: http://codereview.chromium.org/100024