Relanding this:
This patch adds a new kind of tests: browser tests.
They are intended to provide an alternative to the UI tests, hopefully helping
with flackiness.
These tests are in-process browser test.
There are 2 projects to build the tests:
- the browser_test.dll that contains the tests, the gtest framework and the
browser classes.
- the browser_test_launcher.exe.

The launcher loads the browser_test.dll for each test, run the test and then
unloads the DLL. This insures the static initializers are run before each test
and that any atexit handler is invoked properly.

This is Windows only for now.

BUG=9764
TEST=Run browser_test.exe
TBR=maruel

Review URL: http://codereview.chromium.org/99097