Merge r14309 and r14310.
This is a modified merge. The original CL includes strings (1 Tab, % Tabs)
that didn't get included in the final translation drop.

I removed the Tab string and re-inserted the () around the tab icons.

'Window' is replaced by the window icon.

Overall, it's not as nice as the final will be, but nicer than just having
the word 'Window' in the list.

R=sky
BUG= 9022
TEST= With more than one window open close a window. Go to the new tab
  page and make sure you see an entry for the closed window. The entry
  should have a chrome-looking window icon and a list of favicons for 
  each tab in that window.

Review URL: http://codereview.chromium.org/99270