Merge r15125
Add chrome/chromium icons as PNGs.

Also, add support to NativeButton for rendering as default independently of behaving as default. This will be used by the default browser infobar.
Adds methods to the infobar delegate interface to allow the delegate to specify that it wants its Accept button to appear as the default button.

http://crbug.com/9049