Merge r14995, r15029, and r15031 to the 2.0 branch.
Fix for issue http://code.google.com/p/chromium/issues/detail?id=3285.
A NPObject can be called as a construct if it implements NPN_Construct, and NOT implements
NPN_InvokeDefault. Otherwise, NPN_InvokeDefault is called even when the object is called
as a constructor.

BUG= 3285
TBR= fqian

Review URL: http://codereview.chromium.org/100312