Make the width of the reset options confirmation dialog localizable.
It's currently hard-coded to be 400, but it needs to be localizable.
I made it localizable and set the value to 80 (which makes it exactly
the same width as the current one) for most locales except for 
Malayalam for which it's too narrow. 

Once reviewed and landed in the trunk, this need to be merged to the 
branch. 

BUG=11267
TEST=Run CHrome with '--lang=ml', go to Options | Advanced and click 
on 'Restore the defaul settings' button at the bottom. The dialog box 
popping up should not have any truncated strings due to the dimension 
of the dialog being too small

Review URL: http://codereview.chromium.org/115011