Fix for b/1837411 -- Interstitial title lines bleed together.
The line height for titles in the warning interstitials is too small. 
In English, a second line starts to abut the line above it if the 
browser is narrow.

In languages with ligatures above the line, a second line bleeds 
into the first, impairing readability. This is compounded in languages
like Tamil and Marathi, where the title of the interstitial warning 
is quite long and likely to break on two lines.

My solution is to set the line height to 140% in the .title.

Review URL: http://codereview.chromium.org/113139