Merge r16015.
Ensure that IO thread is not called in last actions by SDCH shutdown

Unregister the sdch_dictionary_fetcher so that it destroys any pending
URLFetcher instances when we are nearing shutdown time.  Dictionaries
are only *potentially* useful (and not needed) for future SDCH 
fetches, so abandoning any pending items is safe and harmless.

Bug= 9669
TBR= jar
Review URL: http://codereview.chromium.org/115436