Use av_rescale_q() for converting FFmpeg timestamps to base::TimeDelta.
Previously we were using integer math to convert to microseconds, but depending on the frame rate and packet size we could introduce enough error that could accumulate and introduce audio/video synchronization drift.  av_rescale_q() is a simple function but is designed to minimize error as much as possible.

Review URL: http://codereview.chromium.org/113598