Facilitate a FieldTrial in the renderer
I added a command line for the renderer that accepts a FieldTrial
name and value, and forces that value to be activated in the
renderer.  As a result, any FieldTrial setting that is specified
by the browser process can be set (forced) in the renderer
process.  Such settings can then be used to establish names
of histograms, which means all processes can work in sync
on a single field trial (and generate data).  This should
allow A/B tests to be run that modulate the page load times.

Dave: Please review/confirm that you are happy with the changes to
render_view.cc.  Note that all I did was change the names and limits
for the histograms (they now go up to 3 minutes).  The MakeName()
allows me to get an A/B test of the impact of DNS pre-resolution.

Mike: Please review the code for passing along switch settings.

r=davemoore,mbelshe
Review URL: http://codereview.chromium.org/115525