Fix 12729: Find box doesn't unregister Esc accelerator.
This regressed recently during porting. When
comparing to see who is registered we need to 
be comparing against a class that implements
FocusChangeListener (and registers as such
with AddFocusChangeListener). In this case,
host_ does not implement that interface and 
doesn't register as a listener, so it will 
never be registered. We should be comparing
using 'this', not 'host_'.

I've fixed this and added a browser_test to 
catch this in the future.

BUG=12729
TEST=Open Find, press Esc, press Esc again 
and the Find box should not briefly appear
and hide again. Also, make sure Esc works 
to cancel navigations.

Review URL: http://codereview.chromium.org/115832