linux: simplify gfx::Font constructor logic and add CHECK()s.
Trying to track down a bug that a user is seeing.
The getter we were using to extract the value before was for
a debugger-compatible string.  Perhaps it was the quote-removal
code that was getting us before; if not, these CHECKs will track
it down.

BUG=12530

Review URL: http://codereview.chromium.org/115877