This CL fixes issue 6223 - Drag and drop a new tab from RTL Chrome: the docked window usually showing out of current screen
The problem is that the dragged tab view is not rendered correctly in RTL locale. The view should be rendered in a position that the mouse is on tabstrip, which is consistent with the position in which the view will be docked when mouse release.

The root cause is the dragged tab view's width is calculated incorrectly in RTL locale.

BUG=http://crbug.com/6223
TEST=Open Chrome in Hebrew, drag and drop a tab, observe the tab view while dragging, the view should be rendered in a position that the mouse is on tabstrip.
Review URL: http://codereview.chromium.org/115938