Form Autofil Options should be a radio group
The Form Autofil is currently just a checkbox. To fit the chomium design, we prefer radios for single questions because it a bit clearer which choice you choose. As well, we merged import data and clear browsing data into browsing data since the page was getting a bit long.

BUG=13222 (http://crbug.com/13222)

TEST=Tested the options pane, the form autofil worked as expected.

Review URL: http://codereview.chromium.org/119146