Issue 8838: Close box not highlighted
We optimize mouse-mouve event handling by ignoring two mouse-move events in a row
which works in most cases except that when we animate tab-closing on windows we
send a duplicate mouse-move event on purpose to highlight the button. Unfortunately
our optimization kicks in.

Fixed by reseting the flag to optimize this before sending the extra mouse-move.

BUG=8838

Review URL: http://codereview.chromium.org/118454