Fix javascript-backslash.html layout test failure.
The fix is to break WebURLLoaderImpl into two classes.  It now has a
reference counted inner class that is the ResourceLoaderBridge::Peer.
This change allows the inner class to stick around after the loader
has been destroyed.  That is necessary since the Peer expects to still
receive OnCompletedRequest after it calls Cancel on the bridge.

R=dglazkov
TEST=covered by layout test
BUG=13786

Review URL: http://codereview.chromium.org/125002