Fix a crash that happens when changing themes.
We were holding references to images that got deleted when themes
changed.  These all happen to be in a NineBox, so have NineBox
reload images when the theme change notification is sent.

Also fix an expose bug that wasn't noticeable in the original
theme.  We need to always place the image at 0, 0 for the
background and paint it all the way across the window (should
get clipped by cairo).

BUG=15366

Review URL: http://codereview.chromium.org/149102