linux: make layout test plugin build and run in test_shell
I used the upstream implementation as a guide for the changes I made,
so style faults are theirs.

Also includes updates to test_expectations for the new tests we pass,
and one new baseline for a test that had an output image that we match
Windows on aside from antialiasing.

BUG=11516

Review URL: http://codereview.chromium.org/150144