linux: fix a layout test by having our test plugin provide two mimetypes
The layout test LayoutTests/fast/js/navigator-mimeTypes-length.html asserts
that the number of mimetypes handled by plugins should be greater than the
number of plugins.  This isn't true if the layout test plugin is the only
plugin and it only handles one mimetype, so we make it handle an additional
fake mimetype.

Review URL: http://codereview.chromium.org/149146