Create a project for maintaining default extensions.
This project is responsible for listing and copying the 
extensions we want to ship with Chrome by default. It 
copies the .crx files listed in the project plus a
.json file from:
  src\chrome\browser\extensions\default_extensions
to
  debug\extensions

... where the mini_installer will pick them up. At the
moment only the json file is getting copied, but we can
add .crx files when they are ready.

BUG=None
TEST=I will give a build to QA for testing this.

Review URL: http://codereview.chromium.org/155016