Check in the original icu 4.2.1  with 6 unused directories removed and header files in source/{common,i18n}/unicode) moved to public/{common,i18n}/unicode.
On top of this tree, I'll port and apply our local changes applied to
icu38 that are still necessary for ICU 4.2.1. This is checked in now to
have our local changes tracked not just with patch files I'll check in
but also with the SVN.

In addition, our build files will be updated. Once everything is ready,
the main DEPS file will be updated to point to icu42 instead of icu38.

Review URL: http://codereview.chromium.org/150066