linux: implement the sad plugin
This required moving the sad plugin resource out of the Windows-specific code.

Note that it doesn't display right yet, since the place to display is covered
by the GtkSocket that used to host the plugin.  But that is a separate issue.
(I had it draw to the side to verify the image was correctly getting loaded.)

BUG=16158

Review URL: http://codereview.chromium.org/155262