Remove reference counting from WebWidgetDelegate.
This was added as a speculative crash fix by eseidel back in June of 2008.  See
http://b/1211911 for details.  It didn't appear to resolve the crash then (as
another patch was required to fix the bug), and it doesn't seem to be necessary
since the RenderView, which is the delegate and owns the WebView, is the one in
charge of allocating the WebView and destroying the WebView.  So, it is not
possible for the WebView to outlive the RenderView.

BUG=16234
TEST=none
R=brettw

Review URL: http://codereview.chromium.org/155292