linux/mac: Fix return status of OpenChannelToPlugin message in case of failure
Currently on linux if the plugin process dies, the PluginProcessHost is not destroyed.
A side effect of that is that if a renderer tries to create a new plugin, the OpenChannelToPlugin message returns success with an empty handle, which is interpreted as the host channel, to which the renderer tries to connect to a second time, causing DCHECKs and havok in the browser process.
This CL fixes that particular problem, returning failure if the browser can't connect to the plugin process.

Review URL: http://codereview.chromium.org/155313