Remove flakiness-inducing part of AutomationProxyTest.NavigateToURLWithTimeout1.
It was loading an URL with big timeout and expecting it not to timeout (ok-ish), but
then it loaded the *same* URL with a small timeout and expecting it to timeout.

It turns out that reloading same URL can be really quick, especially on Linux.

And we're testing the scenario when it doesn't timeout in NavigateToURLWithTimeout2.
It just doesn't "preload" the page.

TEST=none
BUG=none

Review URL: http://codereview.chromium.org/149461