Merge 20436 - Fix a crash that happens if a tab is closed while we're in the middle of a drag originating from the tab.
The problem is that the tab gets deleted out from under the
drag operation which is happening in a nested message loop.

To work around this, if we're in the middle of a drag and
we get a tab close request, delay the tab close until after
the drag operation is finished.

BUG=16280

Review URL: http://codereview.chromium.org/149466

TBR=tc@google.com

Review URL: http://codereview.chromium.org/155548