Make sure installer_util_string is always built before installer_util. It was triggering a bug in Incredibuild.
By making one of the dependent of install_util a dummy_executable,
and making this dummy_executable depends on installer_utils_strings,
the odds that it triggers a bug is a lot smaller. I tried 6 clobber
builds so far and none of them failed.

Review URL: http://codereview.chromium.org/155923