Disk cache: Fix handling of invalid entries that are detected when doing evictions or enumerations.
This cl fixes an issue with dirty entries being deleted
twice from disk, and improves the whole thing so that
different flavors or corrupt entries are removed
gracefully. We started deleting things twice when we
created the map of currently-open entries, because
we don't want to have dirty entries on that map, so it is
possible to have multiple EntryImpl objects that refer
to the same entry (for corrupt entries).

BUG=17474
TEST=Unittests, stress_cache.exe

Review URL: http://codereview.chromium.org/155951