Implement delete key for the Remove action in the bookmark manager.
Although we will also need to implement certain other hotkeys such as Ctrl+C and so forth, I see no point in adding an accelerator group as we do not want the accelerators to work when the tree views don't have focus. Also we don't want the menu to show the accelerators.

Also, fix a bug where the organize menu could be stale. mdm pointed out that ctrl+clicking the selection could lead to an empty selection.

BUG=17393

Review URL: http://codereview.chromium.org/160137