Merge 21417 - Add support to URLRequest for deferring redirects.
I chose to add an out parameter to OnReceivedRedirect because it allows for the
default behavior to remain the same.

I considered adding a ContinueAfterRedirect method that all OnReceivedRedirect
implementations would need to call, but this caused one annoying problem:  In
the case of a ChromePlugin, it is possible for the URLRequest to get deleted
inside the handler for the redirect.  This would make it hard to subsequently
call a method on the URLRequest since I would need to have a way to determine
if the URLRequest had been deleted.

TEST=covered by unit tests
BUG=16413,6442
R=eroman,wtc

Review URL: http://codereview.chromium.org/155897

TBR=darin@chromium.org

Review URL: http://codereview.chromium.org/159441